msm: vidc: Amend video scaling support check condition
Currently video driver returns error to the client if video
hardware does not support video scaling, but video driver is
not expected to return error when client has not requested for
video scaling. Amend the check condition to return error only
when client requested for video scaling and video hardware
does not support scaling.

Signed-off-by: Maheshwar Ajja <majja@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>