msm: vidc: Error handling when core is in bad state
When FW is in bad state, it will not respond to any of the driver's
commands. So all the state transitions will time out. Driver needs
to handle these timeouts properly. When core is in bad state, driver
still needs to free OCMEM. If not, video driver leaks OCMEM clock.

Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>