soc: qcom: smd: Range check edge on open
The edge passed into smd_named_open_on_edge() is used as an array index
to see if the requested edge is available.  If the edge value exceeds the
defined set of edges, an access passed the end of the array would happen
which may cause a crash.  Fix this by range checking the edge value before
use.

Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>