thermal: tsens: Fix TSENS Upper/Lower IRQ type
Update TSENS IRQ to level trigger instead of using
edge trigger interrupt.

When a sensors temperature threshold is crossed the
controller asserts a level trigger. This level goes
low when the Upper/Lower status threshold is masked.
However, if another sensor threshold is crossed while we
are in the middle of servicing the previous sensors
threshold crossing in the workqueue the new sensors
threshold crossing will be missed as the interrupt level
remains high. Fix it by registering the interrupt as a
level interrupt, disable the interrupt in the
handler and enable it while we exit from the workqueue.

Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>